+++ pfscf/content/rectangle.go
@@ -7,6 +7,7 @@ import (
 	"strings"
 
 	"github.com/Blesmol/pfscf/pfscf/args"
+	"github.com/Blesmol/pfscf/pfscf/canvas"
 	"github.com/Blesmol/pfscf/pfscf/param"
 	"github.com/Blesmol/pfscf/pfscf/preset"
 	"github.com/Blesmol/pfscf/pfscf/stamp"
@@ -19,11 +20,12 @@ const (
 
 // rectangle needs a description
 type rectangle struct {
-	X, Y    float64
-	X2, Y2  float64
-	Color   string
-	Opacity float64 // TODO convert to ptr
-	Presets []string
+	X, Y         float64
+	X2, Y2       float64
+	Color        string
+	Transparency float64 // TODO convert to ptr
+	Canvas       string
+	Presets      []string
 }
 
 func newRectangle() *rectangle {
@@ -34,8 +36,8 @@ func newRectangle() *rectangle {
 
 // isValid checks whether the current content object is valid and returns an
 // error with details if the object is not valid.
-func (ce *rectangle) isValid(paramStore *param.Store) (err error) {
-	err = utils.CheckFieldsAreSet(ce, "Color")
+func (ce *rectangle) isValid(paramStore *param.Store, canvasStore *canvas.Store) (err error) {
+	err = utils.CheckFieldsAreSet(ce, "Color", "Canvas")
 	if err != nil {
 		return contentValErr(ce, err)
 	}
@@ -55,12 +57,17 @@ func (ce *rectangle) isValid(paramStore *param.Store) (err error) {
 		return contentValErr(ce, err)
 	}
 
+	if _, exists := canvasStore.Get(ce.Canvas); !exists {
+		err = fmt.Errorf("Canvas '%v' does not exist", ce.Canvas)
+		return contentValErr(ce, err)
+	}
+
 	if _, _, _, err = parseColor(ce.Color); err != nil {
 		return contentValErr(ce, err)
 	}
 
-	if ce.Opacity < 0.0 || ce.Opacity > 1.0 {
-		err = fmt.Errorf("Opacity value outside of range 0.0 to 1.0: %v", ce.Opacity)
+	if ce.Transparency < 0.0 || ce.Transparency > 1.0 {
+		err = fmt.Errorf("Transparency value outside of range 0.0 to 1.0: %v", ce.Transparency)
 		return contentValErr(ce, err)
 	}
 
@@ -77,15 +84,15 @@ func (ce *rectangle) resolve(ps preset.Store) (err error) {
 
 	for _, presetID := range ce.Presets {
 		preset, _ := ps.Get(presetID)
-		if err = fillPublicFieldsFromPreset(ce, &preset, "Presets"); err != nil {
+		if err = preset.FillPublicFieldsFromPreset(ce, "Presets"); err != nil {
 			err = fmt.Errorf("Error resolving content: %v", err)
 			return
 		}
 	}
 
 	// defaults
-	if !utils.IsSet(ce.Opacity) {
-		ce.Opacity = 1.0
+	if !utils.IsSet(ce.Transparency) {
+		ce.Transparency = 0.0
 	}
 
 	return nil
@@ -98,7 +105,8 @@ func (ce *rectangle) generateOutput(s *stamp.Stamp, as *args.Store) (err error) 
 		return err
 	}
 
-	s.DrawRectangle(ce.X, ce.Y, ce.X2, ce.Y2, "F", 0, 0, 0, r, g, b, ce.Opacity)
+	style := stamp.RectStyle{Style: "F", FillR: r, FillG: g, FillB: b, Transparency: ce.Transparency}
+	s.DrawRectangle(ce.Canvas, ce.X, ce.Y, ce.X2, ce.Y2, style)
 
 	return nil
 }
@@ -140,11 +148,13 @@ func parseColor(color string) (r, g, b int, err error) {
 // TODO create generic deep-copy function for public fields
 func (ce *rectangle) deepCopy() Entry {
 	copy := rectangle{
-		X:     ce.X,
-		Y:     ce.Y,
-		X2:    ce.X2,
-		Y2:    ce.Y2,
-		Color: ce.Color,
+		X:            ce.X,
+		Y:            ce.Y,
+		X2:           ce.X2,
+		Y2:           ce.Y2,
+		Color:        ce.Color,
+		Transparency: ce.Transparency,
+		Canvas:       ce.Canvas,
 	}
 	for _, preset := range ce.Presets {
 		copy.Presets = append(copy.Presets, preset)