+++ pfscf/content/textCell.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 
 	"github.com/Blesmol/pfscf/pfscf/args"
+	"github.com/Blesmol/pfscf/pfscf/canvas"
 	"github.com/Blesmol/pfscf/pfscf/param"
 	"github.com/Blesmol/pfscf/pfscf/preset"
 	"github.com/Blesmol/pfscf/pfscf/stamp"
@@ -23,6 +24,7 @@ type textCell struct {
 	Font     string
 	Fontsize float64
 	Align    string
+	Canvas   string
 	Presets  []string
 }
 
@@ -34,8 +36,8 @@ func newTextCell() *textCell {
 
 // isValid checks whether the current content object is valid and returns an
 // error with details if the object is not valid.
-func (ce *textCell) isValid(paramStore *param.Store) (err error) {
-	err = utils.CheckFieldsAreSet(ce, "Value", "Font", "Fontsize")
+func (ce *textCell) isValid(paramStore *param.Store, canvasStore *canvas.Store) (err error) {
+	err = utils.CheckFieldsAreSet(ce, "Value", "Font", "Fontsize", "Canvas")
 	if err != nil {
 		return contentValErr(ce, err)
 	}
@@ -55,6 +57,11 @@ func (ce *textCell) isValid(paramStore *param.Store) (err error) {
 		return contentValErr(ce, err)
 	}
 
+	if _, exists := canvasStore.Get(ce.Canvas); !exists {
+		err = fmt.Errorf("Canvas '%v' does not exist", ce.Canvas)
+		return contentValErr(ce, err)
+	}
+
 	return nil
 }
 
@@ -69,7 +76,7 @@ func (ce *textCell) resolve(ps preset.Store) (err error) {
 	// apply presets
 	for _, presetID := range ce.Presets {
 		preset, _ := ps.Get(presetID)
-		if err = fillPublicFieldsFromPreset(ce, &preset, "Presets"); err != nil {
+		if err = preset.FillPublicFieldsFromPreset(ce, "Presets"); err != nil {
 			err = fmt.Errorf("Error resolving content: %v", err)
 			return
 		}
@@ -85,8 +92,8 @@ func (ce *textCell) generateOutput(s *stamp.Stamp, as *args.Store) (err error) {
 		return nil // nothing to do here...
 	}
 
-	y2 := s.DeriveY2(ce.Y, ce.Y2, ce.Fontsize)
-	s.AddTextCell(ce.X, ce.Y, ce.X2, y2, ce.Font, ce.Fontsize, ce.Align, *value, true)
+	y2 := s.DeriveY2(ce.Canvas, ce.Y, ce.Y2, ce.Fontsize)
+	s.AddTextCell(ce.Canvas, ce.X, ce.Y, ce.X2, y2, ce.Font, ce.Fontsize, ce.Align, *value, true)
 
 	return nil
 }